backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.14.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
gunicorn==21.2.0
psycopg[binary]>=3.2.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
reportlab>=4.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0